requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
apache-airflow-providers-amazon==9.1.0
apache-airflow-providers-airbyte==5.2.1
apache-airflow-providers-snowflake==6.5.0
dlt[postgres, snowflake, bigquery, duckdb]==1.14.1
astro-sdk-python
polars
psycopg2-binary